rocr/aie: Moving PDI CU mask metadata in KMQ metadata
